refactor(js): extract pure helpers from creative_row_editor into module
